+++ src/RestlessPiles/PileSettings.cs
@@ -8,5 +8,6 @@ internal static class PileSettings
     internal static IDisposable Register() => SettingsPageApi.RegisterSettings(Plugin.PluginGuid,
         new SettingsSection("Resource piles",
             new SettingsOption("Enable RestlessPiles", PileConfig.Enabled, true),
-            new SettingsOption("Deposit range", PileConfig.Range, true, "m")));
+            new SettingsOption("Isolate pile range", PileConfig.Isolate, true),
+            new SettingsOption("Pile range", PileConfig.Range, true, "m")));
 }